Add in error checking if no CSEM Modular is compiled
This commit ensures that ModEM-CSEM aborts properly if there is no CSEM modular
compiled and if an operation is run on a CSEM data type.

In order to run with a CSEM modular, either -DCSEM_EM1D or -DCSEM_Dipole1D
needs to be defined.

Before this commit, ModEM would appear to make a call to a CSEM modular, but
the resulting gradient would be 0 after the first calculation, as no CSEM 1D
forward would actually be calculated.