Pick default CSEM Modular depending on compiler flags
Before this commit, the EM1D would always be selected when running CSEM even if
CSEM_Dipole1D was defined. This would cause issues when running and compiling
with Dipole1D as no CSEM code would actually be called to compute the 1D
forward, which would make the norm of the gradient 0.

One reason for this is that the ON code developed a new control file schema.
The default CSEM selection for that control file is EM1D.

I have not used much of their control file, mostly for time and not wanting to
futz with it, but this change will respect the CSEM version that is requested
at compile time (via -DCSEM_Dipole1D and -DCSEM_EM1D), preventing the above
problem with Dipole1D.

It should also respect the choice given if a control file is used.