Remove option for Dipole1D+EM1D in configuration file
In order to a bit more user friendly, this commit, along with the commits
previous to this, remove the option to specify Dipole1D+EM1D in the
configuration file.

Users can still specify both in the Makefile, but they will need to ensure that
they use the control file to pick which CSEM library they want to use.