Allow SolnVectors_t and cvectors_t be able to be place holders
This commit allows `solnVectors_t` and `cvectors_t` to be created as 'place holders'.
These 'place holder' `solnVectors_t` and `cvectors_t` will only contain the
information on that electrical solution and wont contain any data.

In conjunction with the `ESolnmManager` this will allow the main task to *not*
allocated eAll, saving a bit of unnecessary memory.