Add routines to allow worker jobs to compute data response
This commit adds in a new subroutine and a new if statement to the worker job
that allows the worker jobs the ability to compute the data response. It
contains code for both the master task and worker tasks.

As of this commits, these routines are not called at all but will be in a
future commit.