Update ESolnManager to take in grid when creating eAll
Apparently, the grid I assigned at the initialization of the ESolnManager is
not what I thought it was. I assumed that this grid target would be updated
throughout the life of a ModEM operation and the EsolnManager could hold it as
a reference for whenever it is needed.

However, this doesn't seem to be the case, for a reason I have not yet dived
into. Apparently, the grid is updated in a different portion of memory than
what I'm pointing too, which was producing partially differing forward results.

This commit matches how the grid was used to create eAll as was done before,
the grid is passed in when eAll is created before each operation.

Not doing this led to some interesting issues. Firstly, the result of the
forward would be wildly different, but only specific components. From my
investigation, calculating the electric field solution was correct, but the
error was being created in the data response calculation, specifically in
`EinterpSetup`. Which was a bit mysterious to debug...

But regardless, this resolves that issue. I don't think this behavior is a
*bug*, but potentially that original memory location is getting lost when it
should not.