Add in a single Configuration file which can generate most ModEM Makefiles
"Three solvers for the Elven-kings under the sky, six options for the
dwarf-lords, and one configuration file to bind them in the command line."

- J.R.R. Tolkien (Allegedly)

This commit adds in a single configuration file, Configure, which has the
ability to generate make files for all solvers, compilers and ModEM build
options.

There is already a lot of description in the Configure file itself, so no need
to re-hash it here, but the goal of this single configure file is to make our
lives easier as developers. Rather than making changes to all 30 something
configuration files, we now only need to manage one.

Currently, this configuration file does not have the capability to generate
HIP/Cuda compatible makefiles. Perhaps that enough is a reason to use a
different build system, currently, we are quite limited with building anything
that is not a Fortran module.

Regardless, this should be easier to maintain, and easier for users to
understand.

This script does make some assumptions, by default it will use MPI, rather than
serial, and will build SP2, if the solver method is not specified.