Add in a warning about using the spherical version of ModEM
This commit adds a warning to the configure file when the user specifies the -s
flag to build a spherical Makefile. Later, we will add in a documentation
page/section that gives more info, when we do we can update the commented out
print statement to point to that section.