Update fmkmf.pl to take in specified mod build argument (i.e. -j/-module etc)
Before, fmkmf.pl made assumption about what argument to use for specifying the
directory to contain .mod files after they were generated. This worked fine,
but is preventing a bit more customizability by the configure file.

One problem, is that fmkmf.pl had the ability to choose the wrong argument in
some instances. For instance, when compiling on a cray machine, one is required
to use the `ftn` wrapper for any Fortran compiler (gnu/ifort/cray). With the
new configure tool you can specify that you want to use this compiler alias by
using the FC environment variable:

```bash
$ FC=ftn ./CONFIG/configure Makefile gfortran
```

In the case above, before this commit, fmkmf.pl would see that the compiler's
name is `ftn`, and it would set the module argument to be: `-module $(OBJDIR)`,
which is not valid for Gfortran.

Thus, this change updates fmkmf.pl to take an extra optional argument -mod
which will allow the configure file to specify what argument to use.

Furthermore, this change should be backwards compatible, so old configure files
should retain the same functionality.