Add ability to pass in common compilation compile variables to configure
This change adds the ability for configure to use commonly set environment
variables that are used for compiling: LDFLAGS, FFLAGS, CPPFLAGS, etc. This
will allow users to pass in specific flags incase they need to link or include
libraries that are not found by default.

It should be noted, that fmkmf.pl kinda trashes LDFLAGS when it is passed in as
it adds in -L to whatever is passed into its -lp command, thus you can't set
LDFLAGS to something like: `-L/path/to/lib` as it will be `-L-L/path/to/lib`
after fmkmf.pl is done with it.