Update flags for ifort and gfortran compilers
This commit re-arranges some flags for both the intel and gnu compilers:

First, it removes any specification of architecture from configure (-mSSE2,
-mavx) etc. For usability, it is probably best for us not to make any
assumptions about the users system. Even adding -xHost is probably not good for
usability, as the user may build on a login node for a compute node.

Thus, I have removed such options. The ability to pass in these instructions
are still there, using FFLAGS. We will probably want to add a section into the
documentation on how the user can compiler ModEM with such arguments.

Lastly, it does some cleanup with arguments, ensuring we are not passsing
gfortran arguments to intel.